move cpp trace code to utils module and add to github issue maker
